+++ app/models/project.rb
@@ -1,6 +1,7 @@
 class Project < ApplicationRecord
   has_many :sections, dependent: :destroy
   has_many :notes
+  has_many :contacts, dependent: :destroy
 
   validates :urn, presence: true, numericality: {only_integer: true}
   validates :trust_ukprn, presence: true, numericality: {only_integer: true}